Extract a trait that provides repository()
This trait will be re-used in the Shell class which offers similar
features to Controller.
